requirements.txt
django==5.0.2
django-environ==0.11.2
django-ninja==1.1.0
psycopg[binary]==3.1.18
requests==2.31.0
unidecode==1.3.8
uvicorn==0.27.1
